Changed locking message output route to go through PNLog

+++ iOS/3.4/pubnub/libs/PubNub/Network/PNReachability.m
@@ -126,7 +126,7 @@ PNReachabilityStatus PNReachabilityStatusForFlags(SCNetworkReachabilityFlags fla
 
                 status = PNReachabilityStatusReachableViaWiFi;
 
-                unsigned long flagsForCleanUp = (unsigned long)flags;
+                unsigned int flagsForCleanUp = (unsigned int)flags;
                 PNBitsOff(&flagsForCleanUp, kSCNetworkReachabilityFlagsReachable, kSCNetworkReachabilityFlagsIsDirect,
                                             kSCNetworkReachabilityFlagsIsLocalAddress, 0);
                 flags = (SCNetworkReachabilityFlags)flagsForCleanUp;
